+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased]
+- Moved scaling of token embeddings from encoder into sinusoidal positional encodings
+
 ## [0.1.4] - 2025-02-14
 - Split settings for layer-norm bias, feedforward bias, and attention bias
 - In- and out-projection of token embeddings now share weights